tools/requirements.txt
matplotlib
simpleaudio
numpy
